+++ sdk/monitor/monitor-opentelemetry-exporter/CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Other Changes
 
 - Renamed Customer Statsbeat feature to customer SDK Stats.
+- Update drop.reason values for customer SDK Stats.
 - Update logic setting ai.location.ip to use the microsoft.client.ip value by default.
 
 ## 1.0.0-beta.33 (2025-08-04)